+++ pkg/util/caps/caps.go
@@ -2,6 +2,7 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
+//go:build linux
 // +build linux
 
 // this is copied from https://golang.org/src/syscall/exec_linux_test.go