+++ webapp/javascript/components/CustomDatePicker.tsx
@@ -8,6 +8,8 @@ import Select from '@webapp/ui/Select';
 
 import styles from './CustomDatePicker.module.scss';
 
+import styles from './CustomDatePicker.module.scss';
+
 interface CustomDatePickerProps {
   from: string;
   until: string;